pyproject.toml
[tool.poetry]
name = "board-game-bot"
version = "0.3.1"
description = "Board game recommender bots 🤖"
authors = ["Markus Shepherd <markus@recommend.games>"]
license = "MIT"
readme = "README.md"
homepage = "https://recommend.games/"
repository = "https://gitlab.com/recommend.games/board-game-bot"
keywords = [
    "board games",
    "tabletop games",
    "bot",
    "twitter"
]
classifiers = [
    "Topic :: Games/Entertainment :: Board Games"
]

[tool.poetry.dependencies]
python = "~3.8"

board-game-utils = "<0.2.0"
html2text = "*"
mastodon-py = "*"
pretty-errors = "*"
python-dotenv = "*"
pytility = "*"
requests = "*"
tweepy = "<4.0.0"
urllib3 = "*"

[tool.poetry.dev-dependencies]
black = "*"
ipython = "*"
mypy = "*"
pre-commit = "*"
psutil = "*"
pylint = "*"
twine = "*"
types-requests = "*"
types-urllib3 = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
